requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.12.2
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
whitenoise==6.2.0
